Add some connection logging
This is about time to make at *least* some system testing.

+++ src/config.rs
@@ -1,4 +1,4 @@
 pub struct Config {
-    pub worker_threads: usize,
+    pub worker_threads: u32,
     pub port: String,
 }
